requirements.txt
EasyProcess==1.1
entrypoint2==1.1
keyboard==0.13.5
MouseInfo==0.1.3
mss==9.0.1
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
pytweening==1.0.7
six==1.16.0
